+++ hal/sparcv8leon/Makefile
@@ -1,11 +1,11 @@
 #
-# Makefile for Phoenix-RTOS loader (SPARCV8 LEON3 HAL)
+# Makefile for Phoenix-RTOS loader (SPARCV8 LEON HAL)
 #
 # Copyright 2022 Phoenix Systems
 #
 # %LICENSE%
 #
 
-include hal/sparcv8leon3/gaisler/Makefile
+include hal/sparcv8leon/gaisler/Makefile
 
 OBJS += $(addprefix $(PREFIX_O)hal/$(TARGET_SUFF)/, string.o interrupts.o exceptions.o _interrupts.o _traps.o)
