+++ hal/sparcv8leon/exceptions.c
