+++ hal/sparcv8leon/gaisler/console.c
