+++ hal/sparcv8leon/gaisler/generic/config.h
@@ -21,14 +21,14 @@
 
 #include "generic.h"
 #include "peripherals.h"
-#include "hal/sparcv8leon3/gaisler/gaisler.h"
-#include "hal/sparcv8leon3/gaisler/types.h"
-#include "hal/sparcv8leon3/cpu.h"
+#include "hal/sparcv8leon/gaisler/gaisler.h"
+#include "hal/sparcv8leon/gaisler/types.h"
+#include "hal/sparcv8leon/cpu.h"
 
-#include <phoenix/arch/sparcv8leon3/syspage.h>
+#include <phoenix/arch/sparcv8leon/syspage.h>
 #include <phoenix/syspage.h>
 
-#define PATH_KERNEL "phoenix-sparcv8leon3-generic.elf"
+#define PATH_KERNEL "phoenix-sparcv8leon-generic.elf"
 
 #endif /* __ASSEMBLY__ */
 
@@ -36,7 +36,7 @@
 #define NWINDOWS 8
 
 /* Import platform specific definitions */
-#include "ld/sparcv8leon3-generic.ldt"
+#include "ld/sparcv8leon-generic.ldt"
 
 
 #endif
