+++ hal/sparcv8leon/gaisler/generic/hal.c
