+++ hal/sparcv8leon/gaisler/gr716/hal.c
