+++ hal/sparcv8leon/gaisler/irqamp.c
