+++ hal/sparcv8leon/gaisler/timer.c
