+++ hal/sparcv8leon/string.c
